Commit of all BEHR files in Jan 2014 form
All files in this version are in the form as they were in Jan 2014.
All such files were either present on the group server, Ashley Russell's old computer, or received from A. Russell's personal backup. The only exception is BEHR_NWUS.m in the 'Old_BEHR_vers' folder, which contains additional comments but no substantive changes to code.