Comments added to corner calculation and he5 methods added.
Added a number of methods to read he5 files using the new "h5read" and "h5info" methods.  These are generic methods that allow the user to specify what group to return.  Also created the scripts in the "for_sally" folder to find modis swaths passing through a particular set of lat and lons.