Work towards finishing the wind speed/direction paper.
Changes are mostly in support of figuring out why the psuedo-retrieval and
cloud-unfiltered full retrieval have a domain wide positive increase in AMF.

NEW FILES:
    Utils/subset_BEHR_pixels.m - allows the user to choose a subset of BEHR
pixels that fit certain geographic constraints, including:
        - are within a specified domain
        - are within a specified distance to a lat/lon point
        - are not in an edge row or at too large a VZA
        - which quadrant they are in relative to that lat/lon point.
    This will prompt the user interactively for these values (except the domain)
unless passed an options structure containing these values. It can also be run
in a mode to just create the options structure.

MODIFIED FILES:
    Utils/misc_behr_wind_plots.m:
        - renamed plot_full_apriori to plot_ens_apriori b/c it can now plot an
ensemble of pseudo-retrieval a priori too. Also made it able to multiply by
scattering weights and restrict to standard BEHR pressures only. Also moved the
code to subset which pixels are chosen to a separate function which is set up to
take the subsetting options as a structure. Conversion to shape factor now uses
the terrain pressure of the pixel to calculate the VCD.

        - plot_pseudo_diff_timeser has more difference options and the ability
to plot the a priori VCD.

        - plot_pseudo_apriori can now plot the hybrid-avg retrievals and convert
to shape factor as well as only plot the standard BEHR pressures.

        - full_apriori_avg_diff adjusted to add an overall average change (not
divided into cloud bins).

        - new subfunction: del_shape_vs_vcd, allows scatter plotting of the
relationship between changes in shape factor, concentration (both at each std.
pressure level), a priori VCD, and AMF or histogram or cumulative distribution
plots of the same changes. Can be done over all pixels, subset of pixel, or a
single pixel.  If a single pixel, it will also plot which pixel so that the
location is clear.

        - new subfunction: cat_cloud_frac_bins, takes a structure output from
plot_ens_apriori and concatenates the cloud bins into a single "Overall" bin.

        - new subfunction: parse_apriori_mode, figures out which UTC hours were
used in the BEHR retrieval for a priori and returns a numeric vector. Currently
unused, I thought I'd need it at one point but ended up not needing it.

        - new subfunction: remove_interp_pres, removes the pressure levels in
the a priori profiles that result from interpolation to the ground or cloud
pressure.

        - new subfunction: return_paths, will return the appropriate paths for
each pseudo or full test retrieval. Not all other subfunctions have been
converted to this yet.

    Utils/uncross_pix_corners.m - first, changes which points the algorithm
checks for a y-coordinate for an intercept value to be between if one slope is
infinite, to be on the points defining the non-infinite slope instead of the
infinite one. Conceivably, a non-rectangular pixel could have an intercept be
between the points defining the infinite slope but not actually have its corners
crossed. This should be a rare case.
        The bigger change is that this now checks if points 1 and 2 (and
therefore 3 and 4) are catty-corner as well as 2 and 3.

    WRF_Utils/read_wrf_vars.m - now accepts DEBUG_LEVEL option input allowing
messages to be suppressed if used in a larger script.

    WRF_Utils/unstagger.m - changed the help text to reflect that the second
variable is "dim" and not "V".

    WRF_Utils/wrf_total_cldfra.m - allow DEBUG_LEVEL input to suppress messages
if used non-interactively. Updated error message for missing zlev variable to
correctly say "zlev" instead of "CLDFRA". Added the option to simply return the
total cloud fraction as a variable rather than write it into the netCDF file,
which is MUCH faster. Updated help text to reflect these changes.