New function used in responding to comments: wrf_grid_area.m
wrf_grid_area estimates the area of a grid of WRF cells given WRF
longitude and latitude centers. It uses wrf_grid_corners, which
estimates the corners as the average of the surrounding centers,
so it's accuracy is not as good as if it actually used the proper
map projection, but it seems to be alright.